src/main/java/de/rabea/UserInterface.java
package de.rabea;

public class UserInterface {

    public String displayBoard(Cell[] cells) {
        String boardPrinter = "";
        int i = 0;
        for (Enum cell : cells) {
            i ++;

            if (cell == Cell.EMPTY) {
                boardPrinter += i;
            } else {
                boardPrinter += cell.toString();
            }
            
            boardPrinter += " ";

            if (i%3 == 0) {
                boardPrinter += "\n";
            }
        }
        return boardPrinter;
    }


    public String askForPosition() {
        return "Please select a position for your mark.";
    }
}
